Добавить контекстное меню с редактированием ячейки категорий
