+++ Tracker/CreatingTrackersVC/CreatingTrackerViewController.swift
@@ -1,0 +1,16 @@
+//
+//  CreateTrackerViewController.swift
+//  Tracker
+//
+//  Created by Konstantin Lyashenko on 06.08.2024.
+//
+
+import UIKit
+
+final class CreatingTrackerViewController: BaseTrackerViewController {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
+
