+++ Tracker/ScheduleVC/ScheduleViewController.swift
@@ -1,0 +1,55 @@
+//
+//  ScheduleViewController.swift
+//  Tracker
+//
+//  Created by Konstantin Lyashenko on 11.08.2024.
+//
+
+import UIKit
+
+class ScheduleViewController: BaseTrackerViewController {
+    
+    private lazy var addDoneButton = UIButton(
+        title: "Готово",
+        backgroundColor: .ypBlack,
+        titleColor: .ypWhite,
+        cornerRadius: 20,
+        font: UIFont.systemFont(ofSize: 16),
+        target: self,
+        action: #selector(addDoneButtonAction)
+    )
+    
+    private lazy var stack: UIStackView = {
+        let stack = UIStackView(arrangedSubviews: [
+            tableView,
+            addDoneButton
+        ])
+        stack.axis = .vertical
+        return stack
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setupUI()
+    }
+    
+    private func setupUI() {
+        [stack].forEach {
+            $0.translatesAutoresizingMaskIntoConstraints = false
+            view.addSubview($0)
+        }
+       
+        NSLayoutConstraint.activate([
+            stack.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+            stack.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
+            stack.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
+            stack.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -16),
+            
+            addDoneButton.heightAnchor.constraint(equalToConstant: 60)
+        ])
+    }
+    
+    @objc private func addDoneButtonAction() {
+        dismiss(animated: true)
+    }
+}
