--- Tracker/TrackersVC/TypeTrackersVC/CreatingTrackersVC/CreatingTrackersVC Cells/ButtonCell.swift
@@ -1,20 +1,0 @@
-//
-//  ButtonCell.swift
-//  Tracker
-//
-//  Created by Konstantin Lyashenko on 06.08.2024.
-//
-
-//import UIKit
-//
-//final class ButtonCell: UITableViewCell {
-//    static let reuseIdentifier = "ButtonCell"
-//    
-//    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-//        super.init(style: style, reuseIdentifier: reuseIdentifier)
-//    }
-//    
-//    required init?(coder: NSCoder) {
-//        fatalError("init(coder:) has not been implemented")
-//    }
-//}
