--- Tracker/TrackersVC/TypeTrackersVC/CreatingTrackersVC/ScheduleVC/ScheduleViewController.swift
@@ -1,15 +1,0 @@
-//
-//  ScheduleViewController.swift
-//  Tracker
-//
-//  Created by Konstantin Lyashenko on 11.08.2024.
-//
-
-import UIKit
-
-class ScheduleViewController: BaseTrackerViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
