+++ Tracker/TypeTrackersVC/TypeTrackersViewController.swift
@@ -1,0 +1,14 @@
+//
+//  TypeTrackersViewController.swift
+//  Tracker
+//
+//  Created by Konstantin Lyashenko on 06.08.2024.
+//
+
+import UIKit
+
+final class TypeTrackersViewController: BaseTrackerViewController {
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
